requirements.txt
beautifulsoup4==4.12.3
feedparser==6.0.11
python-dotenv==1.0.1
requests==2.32.3
twilio==9.2.3
